+++ src/main/java/edu/ie3/datamodel/io/source/couchbase/CouchbaseWeatherSource.java
@@ -5,6 +5,8 @@
 */
 package edu.ie3.datamodel.io.source.couchbase;
 
+import static edu.ie3.datamodel.io.naming.FieldNamingStrategy.WEATHER_COORDINATE_ID;
+
 import com.couchbase.client.core.error.DecodingFailureException;
 import com.couchbase.client.core.error.DocumentNotFoundException;
 import com.couchbase.client.java.json.JsonObject;
@@ -12,6 +14,7 @@ import com.couchbase.client.java.kv.GetResult;
 import com.couchbase.client.java.query.QueryResult;
 import edu.ie3.datamodel.exceptions.NoDataException;
 import edu.ie3.datamodel.exceptions.SourceException;
+import edu.ie3.datamodel.exceptions.ValidationException;
 import edu.ie3.datamodel.io.connectors.CouchbaseConnector;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedWeatherValueData;
 import edu.ie3.datamodel.io.factory.timeseries.TimeBasedWeatherValueFactory;
@@ -96,8 +99,8 @@ public class CouchbaseWeatherSource extends WeatherSource {
   }
 
   @Override
-  public Optional<Set<String>> getSourceFields() {
-    return connector.getSourceFields();
+  public void validate() throws ValidationException {
+    validate(getInputClass(), connector::getSourceFields);
   }
 
   @Override
@@ -305,7 +308,7 @@ public class CouchbaseWeatherSource extends WeatherSource {
           jsonWeatherInputs.stream()
               .map(
                   json -> {
-                    int coordinateId = json.getInt(COORDINATE_ID);
+                    int coordinateId = json.getInt(WEATHER_COORDINATE_ID.toLowerCase());
                     Optional<Point> coordinate = idCoordinateSource.getCoordinate(coordinateId);
                     ZonedDateTime timestamp =
                         weatherFactory.toZonedDateTime(