+++ src/main/java/edu/ie3/datamodel/io/source/csv/CsvWeatherSource.java
@@ -17,6 +17,7 @@ import edu.ie3.datamodel.io.factory.timeseries.TimeBasedWeatherValueFactory;
 import edu.ie3.datamodel.io.naming.FileNamingStrategy;
 import edu.ie3.datamodel.io.naming.timeseries.ColumnScheme;
 import edu.ie3.datamodel.io.naming.timeseries.FileIndividualTimeSeriesMetaInformation;
+import edu.ie3.datamodel.io.source.DataSource;
 import edu.ie3.datamodel.io.source.IdCoordinateSource;
 import edu.ie3.datamodel.io.source.WeatherSource;
 import edu.ie3.datamodel.models.Entity;
@@ -27,7 +28,7 @@ import edu.ie3.datamodel.models.value.WeatherValue;
 import edu.ie3.datamodel.utils.ExceptionUtils;
 import edu.ie3.datamodel.utils.TimeSeriesUtils;
 import edu.ie3.datamodel.utils.Try;
-import edu.ie3.datamodel.utils.Try.*;
+import edu.ie3.datamodel.utils.Try.Failure;
 import edu.ie3.util.interval.ClosedInterval;
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
@@ -42,6 +43,7 @@ import org.locationtech.jts.geom.Point;
 
 /** Implements a WeatherSource for CSV files by using the CsvTimeSeriesSource as a base */
 public class CsvWeatherSource extends WeatherSource {
+  private Set<String> headlineFields = Collections.emptySet();
   private final Map<Point, IndividualTimeSeries<WeatherValue>> coordinateToTimeSeries;
 
   private final CsvDataSource dataSource;
@@ -71,22 +73,9 @@ public class CsvWeatherSource extends WeatherSource {
 
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
 
-  /** Returns an empty optional for now. */
   @Override
-  public Optional<Set<String>> getSourceFields() {
-    return dataSource
-        .getCsvIndividualTimeSeriesMetaInformation(ColumnScheme.WEATHER)
-        .values()
-        .stream()
-        .findFirst()
-        .flatMap(
-            meta -> {
-              try {
-                return dataSource.getSourceFields(meta.getFullFilePath());
-              } catch (SourceException e) {
-                return Optional.empty();
-              }
-            });
+  public void validate() throws ValidationException {
+    DataSource.validate(headlineFields, getInputClass());
   }
 
   @Override
@@ -331,8 +320,9 @@ public class CsvWeatherSource extends WeatherSource {
     try (BufferedReader reader = bufferedReader) {
       final String[] headline = dataSource.parseCsvRow(reader.readLine(), dataSource.csvSep);
 
+      this.headlineFields = Set.of(headline);
       // validating read file
-      weatherFactory.validate(Set.of(headline), WeatherValue.class).getOrThrow();
+      validate();
 
       // by default try-with-resources closes the reader directly when we leave this method (which
       // is wanted to avoid a lock on the file), but this causes a closing of the stream as well.